+++ lib/src/model/sketch/point/point.freezed.dart
@@ -21,16 +21,15 @@ Point _$PointFromJson(Map<String, dynamic> json) {
 class _$PointTearOff {
   const _$PointTearOff();
 
-  _Point call(double x, double y, {required int time, double pressure = 1}) {
+  _Point call(double x, double y, {double pressure = 1}) {
     return _Point(
       x,
       y,
-      time: time,
       pressure: pressure,
     );
   }
 
-  Point fromJson(Map<String, Object> json) {
+  Point fromJson(Map<String, Object?> json) {
     return Point.fromJson(json);
   }
 }
@@ -42,7 +41,6 @@ const $Point = _$PointTearOff();
 mixin _$Point {
   double get x => throw _privateConstructorUsedError;
   double get y => throw _privateConstructorUsedError;
-  int get time => throw _privateConstructorUsedError;
   double get pressure => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
@@ -54,7 +52,7 @@ mixin _$Point {
 abstract class $PointCopyWith<$Res> {
   factory $PointCopyWith(Point value, $Res Function(Point) then) =
       _$PointCopyWithImpl<$Res>;
-  $Res call({double x, double y, int time, double pressure});
+  $Res call({double x, double y, double pressure});
 }
 
 /// @nodoc
@@ -69,7 +67,6 @@ class _$PointCopyWithImpl<$Res> implements $PointCopyWith<$Res> {
   $Res call({
     Object? x = freezed,
     Object? y = freezed,
-    Object? time = freezed,
     Object? pressure = freezed,
   }) {
     return _then(_value.copyWith(
@@ -81,10 +78,6 @@ class _$PointCopyWithImpl<$Res> implements $PointCopyWith<$Res> {
           ? _value.y
           : y // ignore: cast_nullable_to_non_nullable
               as double,
-      time: time == freezed
-          ? _value.time
-          : time // ignore: cast_nullable_to_non_nullable
-              as int,
       pressure: pressure == freezed
           ? _value.pressure
           : pressure // ignore: cast_nullable_to_non_nullable
@@ -98,7 +91,7 @@ abstract class _$PointCopyWith<$Res> implements $PointCopyWith<$Res> {
   factory _$PointCopyWith(_Point value, $Res Function(_Point) then) =
       __$PointCopyWithImpl<$Res>;
   @override
-  $Res call({double x, double y, int time, double pressure});
+  $Res call({double x, double y, double pressure});
 }
 
 /// @nodoc
@@ -114,7 +107,6 @@ class __$PointCopyWithImpl<$Res> extends _$PointCopyWithImpl<$Res>
   $Res call({
     Object? x = freezed,
     Object? y = freezed,
-    Object? time = freezed,
     Object? pressure = freezed,
   }) {
     return _then(_Point(
@@ -126,10 +118,6 @@ class __$PointCopyWithImpl<$Res> extends _$PointCopyWithImpl<$Res>
           ? _value.y
           : y // ignore: cast_nullable_to_non_nullable
               as double,
-      time: time == freezed
-          ? _value.time
-          : time // ignore: cast_nullable_to_non_nullable
-              as int,
       pressure: pressure == freezed
           ? _value.pressure
           : pressure // ignore: cast_nullable_to_non_nullable
@@ -141,8 +129,7 @@ class __$PointCopyWithImpl<$Res> extends _$PointCopyWithImpl<$Res>
 /// @nodoc
 @JsonSerializable()
 class _$_Point extends _Point {
-  const _$_Point(this.x, this.y, {required this.time, this.pressure = 1})
-      : super._();
+  const _$_Point(this.x, this.y, {this.pressure = 1}) : super._();
 
   factory _$_Point.fromJson(Map<String, dynamic> json) =>
       _$$_PointFromJson(json);
@@ -151,39 +138,28 @@ class _$_Point extends _Point {
   final double x;
   @override
   final double y;
-  @override
-  final int time;
   @JsonKey(defaultValue: 1)
   @override
   final double pressure;
 
   @override
   String toString() {
-    return 'Point(x: $x, y: $y, time: $time, pressure: $pressure)';
+    return 'Point(x: $x, y: $y, pressure: $pressure)';
   }
 
   @override
   bool operator ==(dynamic other) {
     return identical(this, other) ||
-        (other is _Point &&
-            (identical(other.x, x) ||
-                const DeepCollectionEquality().equals(other.x, x)) &&
-            (identical(other.y, y) ||
-                const DeepCollectionEquality().equals(other.y, y)) &&
-            (identical(other.time, time) ||
-                const DeepCollectionEquality().equals(other.time, time)) &&
+        (other.runtimeType == runtimeType &&
+            other is _Point &&
+            (identical(other.x, x) || other.x == x) &&
+            (identical(other.y, y) || other.y == y) &&
             (identical(other.pressure, pressure) ||
-                const DeepCollectionEquality()
-                    .equals(other.pressure, pressure)));
+                other.pressure == pressure));
   }
 
   @override
-  int get hashCode =>
-      runtimeType.hashCode ^
-      const DeepCollectionEquality().hash(x) ^
-      const DeepCollectionEquality().hash(y) ^
-      const DeepCollectionEquality().hash(time) ^
-      const DeepCollectionEquality().hash(pressure);
+  int get hashCode => Object.hash(runtimeType, x, y, pressure);
 
   @JsonKey(ignore: true)
   @override
@@ -197,20 +173,17 @@ class _$_Point extends _Point {
 }
 
 abstract class _Point extends Point {
-  const factory _Point(double x, double y,
-      {required int time, double pressure}) = _$_Point;
+  const factory _Point(double x, double y, {double pressure}) = _$_Point;
   const _Point._() : super._();
 
   factory _Point.fromJson(Map<String, dynamic> json) = _$_Point.fromJson;
 
   @override
-  double get x => throw _privateConstructorUsedError;
+  double get x;
   @override
-  double get y => throw _privateConstructorUsedError;
+  double get y;
   @override
-  int get time => throw _privateConstructorUsedError;
-  @override
-  double get pressure => throw _privateConstructorUsedError;
+  double get pressure;
   @override
   @JsonKey(ignore: true)
   _$PointCopyWith<_Point> get copyWith => throw _privateConstructorUsedError;