+++ lib/src/scribble.notifier.dart
@@ -1,7 +1,9 @@
-import 'dart:math';
+import 'dart:typed_data';
+import 'dart:ui' as ui;
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 import 'package:history_state_notifier/history_state_notifier.dart';
 import 'package:scribble/src/model/sketch/sketch.dart';
@@ -10,6 +12,10 @@ import 'package:state_notifier/state_notifier.dart';
 
 abstract class ScribbleNotifierBase extends StateNotifier<ScribbleState> {
   ScribbleNotifierBase(ScribbleState state) : super(state);
+
+  /// You need to provide a key that the [RepointBoundary] can use so you can
+  /// access it from the [renderImage] method.
+  GlobalKey get repaintBoundaryKey;
 
   void onPointerHover(PointerHoverEvent event);
 
@@ -22,16 +28,39 @@ abstract class ScribbleNotifierBase extends StateNotifier<ScribbleState> {
   void onPointerCancel(PointerCancelEvent event);
 
   void onPointerExit(PointerExitEvent event);
+
+  /// Used to render the image to ByteData which can then be stored or reused
+  /// for example in an [Image.memory] widget.
+  ///
+  /// Use [pixelRatio] to increase the resolution of the resulting image.
+  /// You can specify a different [format], by default this method
+  /// generates pngs.
+  Future<ByteData> renderImage({
+    double pixelRatio = 1.0,
+    ui.ImageByteFormat format = ui.ImageByteFormat.png,
+  }) async {
+    final RenderRepaintBoundary? renderObject =
+        repaintBoundaryKey.currentContext?.findRenderObject()
+            as RenderRepaintBoundary?;
+    if (renderObject == null) {
+      throw StateError(
+          "Tried to convert Scribble to Image, but no valid RenderObject was found!");
+    }
+    final img = await renderObject.toImage(pixelRatio: pixelRatio);
+    return (await img.toByteData(format: format))!;
+  }
 }
 
 /// This class controls the state and behavior for a [Scribble] widget.
-class ScribbleNotifier extends StateNotifier<ScribbleState>
-    with HistoryStateNotifierMixin<ScribbleState>
-    implements ScribbleNotifierBase {
+class ScribbleNotifier extends ScribbleNotifierBase
+    with HistoryStateNotifierMixin<ScribbleState> {
   ScribbleNotifier({
     /// If you pass a sketch here, the notifier will use that sketch as a
     /// starting point.
     Sketch? sketch,
+
+    /// Which pointers can be drawn with and are captured.
+    ScribblePointerMode allowedPointersMode = ScribblePointerMode.all,
 
     /// How many states you want stored in the undo history, 30 by default.
     int maxHistoryLength = 30,
@@ -45,14 +74,16 @@ class ScribbleNotifier extends StateNotifier<ScribbleState>
     /// recording, by default it's linear.
     this.pressureCurve = Curves.linear,
   }) : super(
-    ScribbleState.drawing(
-      sketch: sketch ?? const Sketch(lines: []),
-      selectedWidth: widths[0],
-    ),
-  ) {
+          ScribbleState.drawing(
+            sketch: sketch ?? const Sketch(lines: []),
+            selectedWidth: widths[0],
+            allowedPointersMode: allowedPointersMode,
+          ),
+        ) {
     state = ScribbleState.drawing(
       sketch: sketch ?? const Sketch(lines: []),
       selectedWidth: widths[0],
+      allowedPointersMode: allowedPointersMode,
     );
     this.maxHistoryLength = maxHistoryLength;
   }
@@ -71,14 +102,35 @@ class ScribbleNotifier extends StateNotifier<ScribbleState>
   /// receive a map.
   Sketch get currentSketch => state.sketch;
 
+  final GlobalKey _repaintBoundaryKey = GlobalKey();
+
+  @override
+  GlobalKey get repaintBoundaryKey => _repaintBoundaryKey;
+
   /// Only apply the sketch from the undo history, otherwise keep current state
   @override
   @protected
-  ScribbleState transformHistoryState(ScribbleState historyState,
-      ScribbleState currentState) {
+  ScribbleState transformHistoryState(
+      ScribbleState historyState, ScribbleState currentState) {
     return currentState.copyWith(
       sketch: historyState.sketch,
     );
+  }
+
+  /// Can be used to update the state of the Sketch externally (e.g. when
+  /// fetching from a server) to what is passed in as [sketch];
+  ///
+  /// Per default, this state of the sketch gets added to the undo history. If
+  /// this is not desired, set [addToUndoHistory] to ``false``.
+  void setSketch({required Sketch sketch, bool addToUndoHistory = true}) {
+    final newState = state.copyWith(
+      sketch: sketch,
+    );
+    if (addToUndoHistory) {
+      state = newState;
+    } else {
+      temporaryState = newState;
+    }
   }
 
   /// Clear the entire drawing.
@@ -105,6 +157,13 @@ class ScribbleNotifier extends StateNotifier<ScribbleState>
     );
   }
 
+  /// Sets the current mode of allowed pointers to the given [ScribblePointerMode]
+  void setAllowedPointersMode(ScribblePointerMode allowedPointersMode) {
+    temporaryState = state.copyWith(
+      allowedPointersMode: allowedPointersMode,
+    );
+  }
+
   /// Sets the zoom factor to allow for adjusting line width.
   ///
   /// If the factor is 2 for example, lines will be drawn half as thick as
@@ -119,47 +178,47 @@ class ScribbleNotifier extends StateNotifier<ScribbleState>
   /// Sets the color of the pen to the given color.
   void setColor(Color color) {
     temporaryState = state.map(
-      drawing: (s) =>
-          ScribbleState.drawing(
-            sketch: s.sketch,
-            selectedColor: color.value,
-            selectedWidth: s.selectedWidth,
-          ),
-      erasing: (s) =>
-          ScribbleState.drawing(
-            sketch: s.sketch,
-            selectedColor: color.value,
-            selectedWidth: s.selectedWidth,
-            scaleFactor: state.scaleFactor,
-            activePointerIds: state.activePointerIds,
-          ),
+      drawing: (s) => ScribbleState.drawing(
+        sketch: s.sketch,
+        selectedColor: color.value,
+        selectedWidth: s.selectedWidth,
+      ),
+      erasing: (s) => ScribbleState.drawing(
+        sketch: s.sketch,
+        selectedColor: color.value,
+        selectedWidth: s.selectedWidth,
+        scaleFactor: state.scaleFactor,
+        activePointerIds: state.activePointerIds,
+      ),
     );
   }
 
   /// Used by the Listener callback to display the pen if desired
   @override
   void onPointerHover(PointerHoverEvent event) {
+    if (!state.supportedPointerKinds.contains(event.kind)) return;
     temporaryState = state.copyWith(
       pointerPosition:
-      event.distance > 10000 ? null : _getPointFromEvent(event),
+          event.distance > 10000 ? null : _getPointFromEvent(event),
     );
   }
 
   /// Used by the Listener callback to start drawing
   @override
   void onPointerDown(PointerDownEvent event) {
+    if (!state.supportedPointerKinds.contains(event.kind)) return;
     ScribbleState s = state;
 
     // Are there already pointers on the screen?
     if (state.activePointerIds.isNotEmpty) {
       s = state.map(
           drawing: (s) =>
-          // If the current line already contains something
-          (s.activeLine != null && s.activeLine!.points.length > 2)
-              ? _finishLineForState(s)
-              : s.copyWith(
-            activeLine: null,
-          ),
+              // If the current line already contains something
+              (s.activeLine != null && s.activeLine!.points.length > 2)
+                  ? _finishLineForState(s)
+                  : s.copyWith(
+                      activeLine: null,
+                    ),
           erasing: (s) => s);
     } else if (state is Drawing) {
       s = (state as Drawing).copyWith(
@@ -179,6 +238,7 @@ class ScribbleNotifier extends StateNotifier<ScribbleState>
   /// Used by the Listener callback to update the drawing
   @override
   void onPointerUpdate(PointerMoveEvent event) {
+    if (!state.supportedPointerKinds.contains(event.kind)) return;
     if (!state.active) {
       temporaryState = state.copyWith(
         pointerPosition: null,
@@ -199,19 +259,20 @@ class ScribbleNotifier extends StateNotifier<ScribbleState>
   /// Used by the Listener callback to finish a line
   @override
   void onPointerUp(PointerUpEvent event) {
+    if (!state.supportedPointerKinds.contains(event.kind)) return;
     final pos =
-    event.kind == PointerDeviceKind.mouse ? state.pointerPosition : null;
+        event.kind == PointerDeviceKind.mouse ? state.pointerPosition : null;
     if (state is Drawing) {
       state = _finishLineForState(_addPoint(event, state)).copyWith(
         pointerPosition: pos,
         activePointerIds:
-        state.activePointerIds.where((id) => id != event.pointer).toList(),
+            state.activePointerIds.where((id) => id != event.pointer).toList(),
       );
     } else if (state is Erasing) {
       state = _erasePoint(event).copyWith(
         pointerPosition: pos,
         activePointerIds:
-        state.activePointerIds.where((id) => id != event.pointer).toList(),
+            state.activePointerIds.where((id) => id != event.pointer).toList(),
       );
     }
   }
@@ -219,27 +280,29 @@ class ScribbleNotifier extends StateNotifier<ScribbleState>
   /// Used by the Listener callback to stop displaying the cursor
   @override
   void onPointerCancel(PointerCancelEvent event) {
+    if (!state.supportedPointerKinds.contains(event.kind)) return;
     if (state is Drawing) {
       state = _finishLineForState(_addPoint(event, state)).copyWith(
         pointerPosition: null,
         activePointerIds:
-        state.activePointerIds.where((id) => id != event.pointer).toList(),
+            state.activePointerIds.where((id) => id != event.pointer).toList(),
       );
     } else if (state is Erasing) {
       state = _erasePoint(event).copyWith(
         pointerPosition: null,
         activePointerIds:
-        state.activePointerIds.where((id) => id != event.pointer).toList(),
+            state.activePointerIds.where((id) => id != event.pointer).toList(),
       );
     }
   }
 
   @override
   void onPointerExit(PointerExitEvent event) {
+    if (!state.supportedPointerKinds.contains(event.kind)) return;
     temporaryState = _finishLineForState(state).copyWith(
       pointerPosition: null,
       activePointerIds:
-      state.activePointerIds.where((id) => id != event.pointer).toList(),
+          state.activePointerIds.where((id) => id != event.pointer).toList(),
     );
   }
 
@@ -264,25 +327,24 @@ class ScribbleNotifier extends StateNotifier<ScribbleState>
   ScribbleState _erasePoint(PointerEvent event) {
     return state.copyWith.sketch(
       lines: state.sketch.lines
-          .where((l) =>
-          l.points.every((p) =>
-          (event.localPosition - p.asOffset).distance >
-              state.selectedWidth))
+          .where((l) => l.points.every((p) =>
+              (event.localPosition - p.asOffset).distance >
+              l.width + state.selectedWidth))
           .toList(),
     );
   }
 
   /// Converts a pointer event to the [Point] on the canvas.
   Point _getPointFromEvent(PointerEvent event) {
-    final p = event.pressureMin == event.pressureMax
+    final overridePressureOnWeb = event is PointerHoverEvent && kIsWeb;
+    final p = overridePressureOnWeb || event.pressureMin == event.pressureMax
         ? 0.5
         : (event.pressure - event.pressureMin) /
-        (event.pressureMax - event.pressureMin);
+            (event.pressureMax - event.pressureMin);
     return Point(
       event.localPosition.dx,
       event.localPosition.dy,
       pressure: pressureCurve.transform(p),
-      time: event.timeStamp.inMilliseconds,
     );
   }